Solution to Check Even or Odd

+++ Check Even or Odd@@6788dd63490637cfc44e1b8d/code.c
@@ -2,8 +2,8 @@
 
 int main(){
 
-    double a;
-    scanf("%lf",a);
+    int a;
+    scanf("%d",a);
 
     if(a%2==0){
         printf("Even");
